# backend/web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/project/create": {
            "post": {
                "description": "client 创建一个新的项目并上传文件, this api makes sure only client can create the project",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "Clinet create a new project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project Field",
                        "name": "field",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project Description",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Clinet Email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Required Skills",
                        "name": "requiredSkills[]",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "upload file",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"Project created successfully\", \"projectId\": 1, \"fileName\": \"filename.pdf\", \"filePath\": \"backend/files/filename.pdf\", \"createdBy\": 1}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"Invalid email\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Client not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"Failed to save file\"} or {\"error\": \"Failed to create project\"} or {\"error\": \"Failed to find or create skill\"} or {\"error\": \"Failed to associate skills\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/project/delete/{projectId}": {
            "delete": {
                "description": "根据项目ID删除项目",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "Delete project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": bool}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "{\"error\": string}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\": string}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/project/detail/{projectId}": {
            "get": {
                "description": "根据项目ID获取项目的详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "Get project detail by projectID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "项目ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ProjectDetailResponse"
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Project not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"Internal Server Error\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/project/get/public_project/list": {
            "get": {
                "description": "is_public 字段 1表示 public， 2 表示未公开, 这里返回的公开的 project 信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "Get pubilic project list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.GetProjectListResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\": string}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/project/modify/{projectId}": {
            "post": {
                "description": "通过 projectId 修改项目详细信息，并更新项目的 client, this api makes sure Projects can only be assigned to clients",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "Modify project detail information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Client Email",
                        "name": "clientEmail",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Required Skills",
                        "name": "requiredSkills[]",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Project Field",
                        "name": "field",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project Description",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Specification File",
                        "name": "spec",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ModifyProjectDetailResponse"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"File not provided\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "{\"error\": \"Project can only be assigned to client\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Project not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\": Internal Error}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/project/preferencedBy/team/{projectId}": {
            "get": {
                "description": "Get the detail of teams that prefer a given project ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project Preference"
                ],
                "summary": "Get teams that prefer a project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.TeamDetailResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Project not found\"}\" or \"{\"error\": \"Teams not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/project/team/allocated/{projectId}": {
            "get": {
                "description": "查看一个 project 被 allocated 的所有 team 的信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project Allocation"
                ],
                "summary": "Get allocated team details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.TeamDetailResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Project not found\"}\" or \"{\"error\": \"Teams not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.AllocatedTeam": {
            "type": "object",
            "properties": {
                "teamId": {
                    "type": "integer"
                },
                "teamName": {
                    "type": "string"
                }
            }
        },
...
        "response.ModifyProjectDetailResponse": {
            "type": "object",
            "properties": {
                "createdBy": {
                    "type": "string"
                },
                "createdByEmail": {
                    "type": "string"
                },
                "createdByUserId": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "response.ProjectDetailResponse": {
            "type": "object",
            "properties": {
                "allocatedTeam": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.AllocatedTeam"
                    }
                },
                "clientEmail": {
                    "type": "string"
                },
                "clientId": {
                    "type": "integer"
                },
                "clientName": {
                    "type": "string"
                },
                "coorEmail": {
                    "type": "string"
                },
                "coorId": {
                    "type": "integer"
                },
                "coorName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "projectId": {
                    "type": "integer"
                },
                "requiredSkills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "specLink": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "tutorEmail": {
                    "type": "string"
                },
                "tutorId": {
                    "type": "integer"
                },
                "tutorName": {
                    "type": "string"
                }
            }
        },
...
    }
}